remove ObjectDetail and rename CollectionDetail to RecordsDetail
