--- src/components/CollectionDetail/styles.scss
@@ -1,2 +1,0 @@
-@import "../../styles/components/collection-detail";
-@import "../../styles/components/accordion";
