--- src/components/ObjectDetail/styles.scss
@@ -1,2 +1,0 @@
-@import "../../styles/components/object-detail";
-@import "../../styles/components/accordion";
