+++ src/components/RecordsDetail/index.js
@@ -33,17 +33,17 @@ const PanelListSection = props =>  (
   )
 
 // TODO: add params to back button href
-const CollectionDetail = ({ collection, isLoading }) => {
+const RecordsDetail = ({ isLoading, records }) => {
   /** Helper function to return text from a note by title */
   const noteText = noteTitle => {
-    let note = collection.notes && collection.notes.filter(n => {return n.title === noteTitle})[0]
+    let note = records.notes && records.notes.filter(n => {return n.title === noteTitle})[0]
     return note ? note.subnotes.map(s => s.content).join("\r\n") : null
   }
 
   /** Boolean indicator for the presence of access and use notes */
   const hasAccessAndUse = collection => {
-    const access = collection.notes && collection.notes.filter(n => {return n.title === "Conditions Governing Access"}).length
-    const use = collection.notes && collection.notes.filter(n => {return n.title === "Conditions Governing Use"}).length
+    const access = records.notes && records.notes.filter(n => {return n.title === "Conditions Governing Access"}).length
+    const use = records.notes && records.notes.filter(n => {return n.title === "Conditions Governing Use"}).length
     return access || use
   }
 
@@ -54,7 +54,7 @@ const CollectionDetail = ({ collection, isLoading }) => {
         <span className="material-icons">keyboard_arrow_left</span>Back to Search
       </a>
     </nav>
-    <h1 className="collection__title">{collection.title || <Skeleton />}</h1>
+    <h1 className="collection__title">{records.title || <Skeleton />}</h1>
     <Accordion className="accordion" preExpanded={["summary"]} allowZeroExpanded={true}>
       <AccordionItem className="accordion__item" uuid="summary">
         <AccordionItemHeading className="accordion__heading" aria-level={2}>
@@ -65,11 +65,11 @@ const CollectionDetail = ({ collection, isLoading }) => {
             (<DetailSkeleton />) :
             (<>
               <PanelListSection
-              title="Creators"
-              listData={collection.creators} />
+                title="Creators"
+                listData={records.creators} />
               <PanelTextSection
                 title="Dates"
-                text={collection.dates && collection.dates.map(d => d.expression).join(", ")} />
+                text={records.dates && records.dates.map(d => d.expression).join(", ")} />
               <PanelTextSection
                 title="Description"
                 text={noteText("Scope and Contents")} />
@@ -78,7 +78,7 @@ const CollectionDetail = ({ collection, isLoading }) => {
             }
         </AccordionItemPanel>
       </AccordionItem>
-      { hasAccessAndUse(collection) ?
+      { hasAccessAndUse(records) ?
         (<AccordionItem className="accordion__item" uuid="accessAndUse">
           <AccordionItemHeading className="accordion__heading" aria-level={2}>
             <AccordionItemButton className="accordion__button">Access and Use</AccordionItemButton>
@@ -93,7 +93,7 @@ const CollectionDetail = ({ collection, isLoading }) => {
           </AccordionItemPanel>
         </AccordionItem>) :
         (null)}
-      { collection.terms && collection.terms.length ?
+      { records.terms && records.terms.length ?
         (<AccordionItem className="accordion__item" uuid="relatedTerms">
             <AccordionItemHeading className="accordion__heading" aria-level={2}>
               <AccordionItemButton className="accordion__button">Related Terms</AccordionItemButton>
@@ -101,7 +101,7 @@ const CollectionDetail = ({ collection, isLoading }) => {
             <AccordionItemPanel className="accordion__panel">
               <PanelListSection
                 title="Subjects"
-                listData={collection.terms} />
+                listData={records.terms} />
             </AccordionItemPanel>
           </AccordionItem>) :
         (null)}
@@ -110,4 +110,4 @@ const CollectionDetail = ({ collection, isLoading }) => {
   )
 }
 
-export default CollectionDetail;
+export default RecordsDetail;
