+++ src/components/RecordsDetail/styles.scss
@@ -1,0 +1,2 @@
+@import "../../styles/components/records-detail";
+@import "../../styles/components/accordion";
