+++ src/styles/components/_records-detail.scss
